+++ lib/backend/src/arch.ml
@@ -12,12 +12,17 @@
 open Base
 
 module M = struct
-  type t = C | Assembly of Act_common.Id.t [@@deriving sexp, compare]
+  type t =
+    | C of {underlying_arch: Act_common.Id.t option}
+    | Assembly of Act_common.Id.t
+  [@@deriving sexp, compare]
 end
 
 include M
 include Comparable.Make (M)
 
-let c : t = C
+let c : t = C {underlying_arch= None}
+
+let c_x86 : t = C {underlying_arch= Some (Act_common.Id.of_string "x86")}
 
 let asm_x86 : t = Assembly (Act_common.Id.of_string "x86")